+++ Gruntfile.js
@@ -38,6 +38,7 @@ module.exports = function (grunt) {
   });
 
   // These plugins provide necessary tasks.
+  grunt.loadTasks('./tasks')
   grunt.loadNpmTasks('grunt-contrib-nodeunit');
   grunt.loadNpmTasks('grunt-contrib-jshint');
   grunt.loadNpmTasks('grunt-contrib-watch');