+++ CHANGELOG.md
@@ -6,6 +6,10 @@ _(none)_
 
 --------------------
 
+## 0.1.4 (2014-01-27)
+* Added line and version args to grunt cli commands
+* Made grunt commands async
+
 ## 0.1.3 (2014-01-27)
 * Added grunt as a dependency